유초원) (수정) follower.jsp, FollowPf, MovieMapper, sidebar -> 팔로워 버튼 구현 중간 커밋

+++ thegrace/src/main/java/com/smhrd/model/FollowPf.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 public class FollowPf {
 	
 	// 팔로우 코드 
-	private Double follow_cd;
+	private int follow_cd;
 	
 	// 팔로우 받은 사람
     private String followee;
